fix(cli): address channel status review feedback
