+++ src/lib/messaging/channels/rendered-config-parser.ts
@@ -1,6 +1,8 @@
 // SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import type { ChannelManifest } from "../manifest";
+import { BUILT_IN_CHANNEL_MANIFESTS } from "./built-ins";
 import { discordRenderedConfigParser } from "./discord/rendered-config-parser";
 import type { RenderedChannelConfigParser } from "./rendered-config-parser-utils";
 import { slackRenderedConfigParser } from "./slack/rendered-config-parser";
@@ -11,17 +13,36 @@ import { whatsappRenderedConfigParser } from "./whatsapp/rendered-config-parser"
 
 export * from "./rendered-config-parser-utils";
 
-const BUILT_IN_RENDERED_CONFIG_PARSERS: Readonly<Record<string, RenderedChannelConfigParser>> = {
-  discord: discordRenderedConfigParser,
-  slack: slackRenderedConfigParser,
-  teams: teamsRenderedConfigParser,
-  telegram: telegramRenderedConfigParser,
-  wechat: wechatRenderedConfigParser,
-  whatsapp: whatsappRenderedConfigParser,
-};
+const BUILT_IN_RENDERED_CONFIG_PARSERS: ReadonlyMap<string, RenderedChannelConfigParser> = new Map(
+  BUILT_IN_CHANNEL_MANIFESTS.map((manifest) => [
+    manifest.id,
+    renderedConfigParserForBuiltInManifest(manifest),
+  ]),
+);
 
 export function getBuiltInRenderedConfigParser(
   channelId: string,
 ): RenderedChannelConfigParser | null {
-  return BUILT_IN_RENDERED_CONFIG_PARSERS[channelId] ?? null;
+  return BUILT_IN_RENDERED_CONFIG_PARSERS.get(channelId) ?? null;
+}
+
+function renderedConfigParserForBuiltInManifest(
+  manifest: ChannelManifest,
+): RenderedChannelConfigParser {
+  switch (manifest.id) {
+    case "discord":
+      return discordRenderedConfigParser;
+    case "slack":
+      return slackRenderedConfigParser;
+    case "teams":
+      return teamsRenderedConfigParser;
+    case "telegram":
+      return telegramRenderedConfigParser;
+    case "wechat":
+      return wechatRenderedConfigParser;
+    case "whatsapp":
+      return whatsappRenderedConfigParser;
+    default:
+      throw new Error(`missing rendered config parser for built-in channel '${manifest.id}'`);
+  }
 }
